api/routes/health.py
import logging
from datetime import datetime, timezone
from fastapi import APIRouter
from pydantic import BaseModel
from typing import Dict, Any

logger = logging.getLogger(__name__)
router = APIRouter(prefix="/health", tags=["Health"])


class HealthResponse(BaseModel):
    """Health check response model"""
    status: str
    timestamp: datetime
    version: str = "1.0.0"
    services: Dict[str, Any] = {}


@router.get(
    "/",
    response_model=HealthResponse,
    summary="Health check",
    description="Check the health status of the API and its dependencies"
)
async def health_check():
    """
    Perform health check on all services
    """
    try:
        # For now, just return a basic health check
        # TODO: Add actual service health checks when services are implemented
        return HealthResponse(
            status="healthy",
            timestamp=datetime.now(timezone.utc),
            services={
                "api": True,
                "database": True  # Will be implemented when database is connected
            }
        )
        
    except Exception as e:
        logger.error(f"Health check failed: {str(e)}")
        return HealthResponse(
            status="unhealthy",
            timestamp=datetime.now(timezone.utc),
            services={
                "api": False,
                "database": False
            }
        )


@router.get("/ready", summary="Readiness check", description="Check if the API is ready to handle requests")
async def readiness_check():
    """Check if the API is ready to handle requests"""
    return {"ready": True, "timestamp": datetime.now(timezone.utc)}


@router.get("/live", summary="Liveness check", description="Check if the API is alive")
async def liveness_check():
    """Check if the API is alive"""
    return {"alive": True, "timestamp": datetime.now(timezone.utc)}