api/requirements.txt
azure-storage-blob
azure-data-tables
pymongo
fastapi
uvicorn
python-dotenv
python-multipart
pillow
pillow-heif
exifread
lxml
pyproj
requests